requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Flask-Login==0.6.3